refactor(server): slim generate.routes through generationService
The biggest route file in the codebase had ~360 lines of orchestration:
voice / model resolution, sample lookup, cache-miss retry with base64
re-upload, three backend branches, pre-create + finalize + cleanup of
the generation record, plus a tee'd-stream save path. Pulls the lot
into generationService:

  generationService.generateBuffered(body, userId)
    Returns { generationId, audio, contentType }. Owns resolve ->
    preCreate -> per-backend run -> finalize, with cleanup on any
    throw so failed generations don't leak placeholder rows.

  generationService.generateStreaming(body, userId)
    Returns either a StreamingGeneration (tee'd inference stream) or
    a BufferedGeneration (cloud backend fallback). The route picks
    the right Response shape from the discriminator.

Resolution now throws typed NotFoundError / BadRequestError instead
of returning hand-rolled `Response.json({...}, 404)`, so the route
just wraps everything in mapServiceError. The cache-miss retry, the
tee + accumulateAndSave background save, and fetchSamplesAsBase64
all live alongside it as private helpers.

Route file drops from 354 lines to 56. Behaviour preserved:
  - same response headers (Content-Type, X-Generation-Id, X-Sample-Rate,
    X-Channels, X-Bits-Per-Sample)
  - same cleanup semantics on inference failure
  - same backend fallback for ElevenLabs / OpenAI on the streaming path
  - same Piper noise_scale derivation from variationSeed
